+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Supports stacking for others
 - Added support for moving position in stack
 - Added links to polciies in footer
+- Added support for canceling a stack
 ### Fix
 - Added client-side form validation to disallow blank room or user names
 - Handles duplicate usernames